order_delivery_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
py-eureka-client==0.12.0
django-cors-headers==4.3.1
requests==2.31.0

